Clean up psi util
