+++ src/nl/hannahsten/texifyidea/completion/TexifyCompletionContributor.kt
@@ -97,7 +97,7 @@ open class TexifyCompletionContributor : CompletionContributor() {
                 .inside(LatexRequiredParam::class.java)
                 .with(object : PatternCondition<PsiElement>("File name completion pattern") {
                     override fun accepts(psiElement: PsiElement, processingContext: ProcessingContext): Boolean {
-                        val command = LatexPsiUtil.getParentOfType(psiElement, LatexCommands::class.java)
+                        val command = getParentOfType(psiElement, LatexCommands::class.java)
                             ?: return false
 
                         val name = command.commandToken.text
@@ -120,7 +120,7 @@ open class TexifyCompletionContributor : CompletionContributor() {
                 .inside(LatexRequiredParam::class.java)
                 .with(object : PatternCondition<PsiElement>("Folder name completion pattern") {
                     override fun accepts(psiElement: PsiElement, processingContext: ProcessingContext): Boolean {
-                        val command = LatexPsiUtil.getParentOfType(psiElement, LatexCommands::class.java)
+                        val command = getParentOfType(psiElement, LatexCommands::class.java)
                             ?: return false
 
                         val name = command.commandToken.text
@@ -143,7 +143,7 @@ open class TexifyCompletionContributor : CompletionContributor() {
                 .inside(LatexRequiredParam::class.java)
                 .with(object : PatternCondition<PsiElement>("Folder name completion pattern") {
                     override fun accepts(psiElement: PsiElement, processingContext: ProcessingContext): Boolean {
-                        val command = LatexPsiUtil.getParentOfType(psiElement, LatexCommands::class.java)
+                        val command = getParentOfType(psiElement, LatexCommands::class.java)
                             ?: return false
 
                         val name = command.commandToken.text
@@ -165,7 +165,7 @@ open class TexifyCompletionContributor : CompletionContributor() {
             PlatformPatterns.psiElement().inside(LatexRequiredParam::class.java)
                 .with(object : PatternCondition<PsiElement>("xcolor color completion patter") {
                     override fun accepts(psiElement: PsiElement, context: ProcessingContext?): Boolean {
-                        val command = LatexPsiUtil.getParentOfType(psiElement, LatexCommands::class.java)
+                        val command = getParentOfType(psiElement, LatexCommands::class.java)
                             ?: return false
 
                         val name = command.commandToken.text
