+++ src/nl/hannahsten/texifyidea/inspections/LatexGrammarCheckingStrategy.kt
@@ -12,21 +12,8 @@ import com.intellij.refactoring.suggested.endOffset
 import com.intellij.refactoring.suggested.startOffset
 import nl.hannahsten.texifyidea.psi.*
 import nl.hannahsten.texifyidea.util.childrenOfType
-
-/**
- * Return true if [this] has a non-empty overlap with [other].
- */
-fun IntRange.overlaps(other: IntRange): Boolean {
-    return this.last >= other.first && other.last >= this.first
-}
-
-/**
- * Merge [this] with all [others] to result in one IntRange, assuming that [this] overlaps with all [others].
- */
-fun IntRange.merge(others: Iterable<IntRange>): IntRange {
-    val all = mutableSetOf(this).also { it.addAll(others) }
-    return IntRange(all.minByOrNull { it.first }!!.first, all.maxByOrNull { it.last }!!.last)
-}
+import nl.hannahsten.texifyidea.util.merge
+import nl.hannahsten.texifyidea.util.overlaps
 
 class LatexGrammarCheckingStrategy : GrammarCheckingStrategy {
