+++ src/nl/hannahsten/texifyidea/inspections/latex/LatexPrimitiveStyleInspection.kt
@@ -12,9 +12,9 @@ import nl.hannahsten.texifyidea.insight.InsightGroup
 import nl.hannahsten.texifyidea.inspections.TexifyInspectionBase
 import nl.hannahsten.texifyidea.psi.LatexCommands
 import nl.hannahsten.texifyidea.psi.LatexNoMathContent
-import nl.hannahsten.texifyidea.psi.LatexPsiUtil.getNextSiblingIgnoreWhitespace
 import nl.hannahsten.texifyidea.util.Magic
 import nl.hannahsten.texifyidea.util.deleteElement
+import nl.hannahsten.texifyidea.util.nextSiblingIgnoreWhitespace
 import nl.hannahsten.texifyidea.util.parentOfType
 import org.jetbrains.annotations.Nls
 import kotlin.reflect.jvm.internal.impl.utils.SmartList
@@ -80,7 +80,7 @@ class LatexPrimitiveStyleInspection : TexifyInspectionBase() {
                 return
             }
             val content = element.parentOfType(LatexNoMathContent::class)
-            val next = getNextSiblingIgnoreWhitespace(content!!)
+            val next = content!!.nextSiblingIgnoreWhitespace()
             val after = if (next == null) "" else next.text
             val replacement =
                 String.format(Magic.Command.stylePrimitveReplacements[cmdIndex], after)
